// robot-project-v3.client/src/modeles/dto.ts
//----------------------
// <auto-generated>
//     Generated using the NSwag toolchain v14.1.0.0 (NJsonSchema v11.0.2.0 (Newtonsoft.Json v13.0.0.0)) (http://NSwag.org)
// </auto-generated>
//----------------------

/* tslint:disable */
/* eslint-disable */

// ReSharper disable InconsistentNaming

export interface ApiResponseError {
  error?: string | undefined;
}

export interface BackTestRequestDto {
  balance?: number;
  minSpread?: number;
  maxSpread?: number;
}

export interface BooleanApiResponse {
  data?: boolean;
}

export interface CandleDto {
  bidVolume?: number;
  askVolume?: number;
  date?: Date;
  open?: number;
  high?: number;
  low?: number;
  close?: number;
  volume?: number;
}

export interface ConnectDto {
  user?: string | undefined;
  pwd?: string | undefined;
  handlerEnum?: string;
}

export interface GlobalResultsDto {
  result?: ResultDto;
  positions?: PositionDto[] | undefined;
  monthlyResults?: MonthlyResultDto[] | undefined;
}

export interface MonthlyResultDto {
  date?: Date;
  result?: ResultDto;
  positions?: PositionDto[] | undefined;
}

export interface PositionDto {
  id?: string | undefined;
  symbol?: string | undefined;
  typePosition?: string | undefined;
  spread?: number | undefined;
  profit?: number | undefined;
  openPrice?: number | undefined;
  dateOpen?: Date;
  closePrice?: number | undefined;
  dateClose?: Date | undefined;
  reasonClosed?: string | undefined;
  stopLoss?: number | undefined;
  takeProfit?: number | undefined;
  volume?: number | undefined;
  pips?: number | undefined;
  statusPosition?: string | undefined;
  comment?: string | undefined;
}

export interface ResultDto {
  drawndownMax?: number;
  drawndown?: number;
  gainMax?: number;
  moyenneNegative?: number;
  moyennePositive?: number;
  moyenneProfit?: number;
  perteMax?: number;
  profit?: number;
  profitFactor?: number;
  profitNegatif?: number;
  profitPositif?: number;
  ratioMoyennePositifNegatif?: number;
  tauxReussite?: number;
  totalPositionNegative?: number;
  totalPositionPositive?: number;
  totalPositions?: number;
}

export interface StrategyCompilationResponseDto {
  compiled?: boolean;
  strategyFileDto?: StrategyFileDto;
  errors?: string[] | undefined;
}

export interface StrategyFileDto {
  id?: number;
  name?: string | undefined;
  version?: string | undefined;
  data?: string | undefined;
  lastDateUpdate?: Date;
}

export interface StrategyInfoDto {
  id: string;
  symbol?: string | undefined;
  strategyName?: string | undefined;
  canRun?: boolean;
  strategyDisabled?: boolean;
  secureControlPosition?: boolean;
  lastTick?: TickDto;
  lastCandle?: CandleDto;
}

export interface StrategyInitDto {
  strategyFileId: string;
  symbol: string;
}

export interface StringApiResponse {
  data?: string | undefined;
}

export interface SymbolInfoDto {
  category?: string | undefined;
  contractSize?: number | undefined;
  currency?: string | undefined;
  currencyProfit?: string | undefined;
  lotMin?: number | undefined;
  precision?: number;
  symbol?: string | undefined;
  tickSize?: number;
  leverage?: number;
}

export interface TickDto {
  ask?: number | undefined;
  askVolume?: number | undefined;
  bid?: number | undefined;
  bidVolume?: number | undefined;
  date?: Date;
  readonly spread?: number | undefined;
}